reccommender-service/requirements.txt
flask>=3.0
pandas>=1.3
numpy>=1.26
sklearn>=1.3
pymongo>=4.6
pyaml>=6.0